Change seats, smseats, smserviceaccounts from short to int like they are in the Bitwarden server code

+++ src/licenseGen/Program.cs
@@ -364,7 +364,7 @@ namespace bitwardenSelfLicensor
             set("Enabled", true);
             set("Plan", "Custom");
             set("PlanType", Enum.Parse(planTypeEnum, "Custom"));
-            set("Seats", (int)short.MaxValue);
+            set("Seats", int.MaxValue);
             set("MaxCollections", short.MaxValue);
             set("UsePolicies", true);
             set("UseSso", true);
@@ -384,8 +384,8 @@ namespace bitwardenSelfLicensor
             set("Issued", DateTime.UtcNow);
             set("UsePasswordManager", true);
             set("UseSecretsManager", true);
-            set("SmSeats", (int)short.MaxValue);
-            set("SmServiceAccounts", (int)short.MaxValue);
+            set("SmSeats", int.MaxValue);
+            set("SmServiceAccounts", int.MaxValue);
             set("Refresh", DateTime.UtcNow.AddYears(100).AddMonths(-1));
             set("Expires", DateTime.UtcNow.AddYears(100));
             set("Trial", false);
